refactor state tests
